Move map to components folder
