+++ src/app/components/map/state/machine.ts
