+++ src/app/components/map/state/types.ts
